scraps の route を分割
